msdk: context: add job type to figure out if joining session is necessary
According to the driver's instruction, if there are two or more encoders
or decoders in a process, the session should be joined by MFXJoinSession.

To achieve this successfully by GstContext, this patch adds job type
specified if it's encoder, decoder or vpp.

If a msdk element gets to know if joining session is needed by the shared context,
it should create another instance of GstContext with joined session, which
is not shared.

https://bugzilla.gnome.org/show_bug.cgi?id=790752